requirements.txt
fastapi
uvicorn
sqlmodel
python-multipart
openai
python-dotenv
aiofiles
requests
passlib[bcrypt]
python-jose[cryptography]
